Implement eject command
