+++ examples/minimal/tsconfig.json
@@ -3,7 +3,7 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "es5",
-    "rootDir": "./src",
+    "rootDir": ".",
     "outDir": "dist",
     "sourceMap": true,
     "lib": ["es2015", "esnext.asynciterable", "es2017", "dom"],
